Fix failed reinterpret_cast during transition
Camera::Thirdperson::UpdateInternalRotation assumes that the current
camera is a in a 3rd person state but when dismounting a horse, we are
in a transition state from the horseback camera to the 3rd person
camera.

This commit fixes the access violation exception we get because we are
trying to interpret and use the state as a CorrectedThirdPersonState
even tho it actually is a CorrectedPlayerCameraTransitionState.

Resolves #54